+++ integrations/flight/sqlite/table_batch_reader.go
